+++ springlesson/src/main/java/com/example/springlesson/service/ProductService.java
@@ -47,7 +47,7 @@ public class ProductService {
      * 商品詳細取得
      */
     @Transactional(readOnly = true)
-    public Product findById(Integer id) {
+    public Product findById(Long id) {
       return productRepository.findById(id.longValue())
               .orElseThrow(() ->
                       new IllegalArgumentException("商品が存在しません。"));
